+++ CHANGES.rst
@@ -4,8 +4,14 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Added `RecreateSessionView.get_create_session_custom_data` to compute
+  `create_session_custom_data` using first element of old session.
+  Moreover, `Recreate session` action is hidden when session is `to_sign`.
+  [gbastien]
+- Added default implementation for `ISignable` adapter so we are sure that every
+  relevant methods have a default implementation.
+  Overrides should inheritates from it.
+  [gbastien]
 
 1.0.1 (2026-08-20)
 ------------------